[android] Extract exceptions to separate files
