+++ packages/expo-font/android/src/main/java/expo/modules/font/exceptions/FontFileInvalidException.java
@@ -1,0 +1,14 @@
+package expo.modules.font.exceptions;
+
+import org.unimodules.core.errors.CodedRuntimeException;
+
+public class FontFileInvalidException extends CodedRuntimeException {
+  public FontFileInvalidException(String path) {
+    super(String.format("File '%s' isn't a valid font file.", path));
+  }
+
+  @Override
+  public String getCode() {
+    return "ERR_FONT_FILE_INVALID";
+  }
+}
